Move spectrum chisqr to chisqr module.
