+++ new_alpha_detect_limit_simulation.py
@@ -11,7 +11,7 @@ from joblib import Parallel, delayed
 
 # self written modules
 from utilities.debug_utils import pv
-from utilities.chisqr import chi_squared
+from utilities.chisqr import spectrum_chisqr
 from models.alpha_model import alpha_model
 from spectrum_overload.Spectrum import Spectrum
 from utilities.simulation_utilities import combine_spectra
@@ -23,41 +23,11 @@ logging.basicConfig(level=logging.DEBUG,
 debug = logging.debug
 
 
-def spectrum_chisqr(spectrum_1, spectrum_2, error=None):
-    """Chi squared for specturm objects."""
-    # Spectrum wrapper for chissquare
-    # make sure xaxis is the Same
-    # if len(spectrum_1) == len(spectrum_2):
-    nan_number = np.sum(np.isnan(spectrum_1.flux))
-    if nan_number:
-        print("There are {} nans in spectrum_1".format(nan_number))
-
-    if np.all(np.isnan(spectrum_2.flux)):
-        print("spectrum 2 is all nans")
-    if np.all(spectrum_1.xaxis == spectrum_2.xaxis):
-        # print("xaxis are equal")
-        c2 = chi_squared(spectrum_1.flux, spectrum_2.flux, error=error)
-        # return chi_squared(spectrum_1.flux, spectrum_2.flux, error=None)
-        # print("chisqrayured value", c2)
-        # spectrum_plotter(spectrum_1, label="obs", show=False)
-        # spectrum_plotter(spectrum_2, label="evauated")
-        if np.isnan(c2):
-            print(" Nan chisqr")
-            # print(spectrum_1.xaxis, spectrum_1.flux, spectrum_2.xaxis, spectrum_2.flux)
-        return c2
-    else:
-
-        print("Spectrum_1", len(spectrum_1))
-        print("Spectrum_2", len(spectrum_2))
-
-        raise Exception("TODO: make xaxis equal in chisquare of spectrum")
-
-
 def model_chisqr_wrapper(spectrum_1, model, params, error=None):
     """Evaluate model and call chisquare."""
     # print("params for model", params)
     # params = copy.copy(params)
-    evaluated_model = model(*params)  # # unpack parameters
+    evaluated_model = model(*params)  # unpack parameters
 
     if np.all(np.isnan(evaluated_model.flux)):
         raise Exception("Evaluated model is all Nans")
